fixed partially the crashing screen error

+++ app/src/main/java/com/example/projettdm/parking_list/data/remote/response/Parking.kt
@@ -3,7 +3,7 @@ package com.example.projettdm.parking_list.data.remote.response
 data class Parking(
     val id: Int,
     val name: String,
-    val price: Int,
+    val price: Float,
     val capacity: Int,
     val available: Int,
     val reserved: Int,
@@ -12,4 +12,6 @@ data class Parking(
     val latitude: Double,
     val longitude: Double,
     val description: String
+//    val setup: String,
+//    val punchline: String,
 )
